remove json support in JSONField for postgresql 9.3, defaults to TextField
'json' field type doesn't play well with django because it doesn't implement
the equality operator, so it cannot be use with DISTINCT clause (ie in admin if
filter_list is set).
The new 'jsonb' in postgresql 9.4 will not have this limitation, so JSONField will
be updated when 9.4 will be out.