set DatabaseCommand.requires_system_checks = False
Sometime apps in setup time try to open a database connection,
and django try to the database command fails.

Behavour spot with django-debug-toolbar which patch urlconf settings,
forcing django-cms to get Sites current site:
if there is no db - as in initdb time - it fails.

Need a better way to prevent the connection, which really is not needed.